noise data + cleanup

+++ ntwewy-db/Classes/NoisepediaEntry.cs
@@ -5,17 +5,17 @@
         public byte Id { get; set; }
         public byte SortIndex { get; set; }
         public int CharacterId { get; set; }
-        public int[] GroupCharacterIds { get; set; }
-        public ushort NoiseId { get; set; }
+        public ushort[] GroupCharacterIds { get; set; }
+        public uint NoiseId { get; set; }
         public string Icon { get; set; }
         public byte SymbolType { get; set; }
         public string Name { get; set; }
         public string Info { get; set; }
         public bool IsBoss { get; set; }
-        public byte[] WeaknessElements { get; set; }
+        public sbyte[] WeaknessElements { get; set; }
         public string NoiseImagePath { get; set; }
 
-        public NoisepediaEntry(byte id, byte sortIndex, int characterId, int[] groupCharacterIds, ushort noiseId, string icon, byte symbolType, string name, string info, bool isBoss, byte[] weaknessElements, string noiseImagePath)
+        public NoisepediaEntry(byte id, byte sortIndex, int characterId, ushort[] groupCharacterIds, uint noiseId, string icon, byte symbolType, string name, string info, bool isBoss, sbyte[] weaknessElements, string noiseImagePath)
         {
             Id = id;
             SortIndex = sortIndex;
